pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = "PythonWeb. HW_8"
authors = [
    {name = "Oksana Donchuk",email = "80112131+OksanaDonchuk@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
sphinx = "^8.2.3"

